Fix duration_ns float->int; clarify the "did not activate" warning
Reported live: "SAO tracing was requested but did not activate for this
run" even after a prior run showed sessions updating in the console. The
warning text only mentioned two of the three failure prints this code
actually has -- it never told the user to look for "SAO trace_call(...)
failed", the one that fires when the tracer builds fine (explaining
sessions appearing) but every individual trace call raises. Fixed the
message to name all three explicitly and ask for the exact line back.

Also fixed while reasoning through possible causes, independent of
whatever the terminal shows: Model._call() computed duration_ns as
`(time.time() - start) * 1e9`, a float, but add_llm_span()'s own
signature types duration_ns as `int | None`. Cast to int() -- correct
regardless of whether this is THE cause of the current failure, since
passing a float where the SDK expects int risks a validation error
inside a call path that doesn't reliably surface such errors back to us.

+++ foundry_mini/observability.py
@@ -99,7 +99,7 @@ class SAOTracer:
         self._session_started = True
 
     def trace_call(self, role: str, system: str, user: str, model_name: str,
-                   reply: str, in_tok: int, out_tok: int, duration_ns: float) -> None:
+                   reply: str, in_tok: int, out_tok: int, duration_ns: int) -> None:
         """One trace + one LLM span for a single Model._call(). Never raises —
         a bad SAO account degrades to 'no tracing', not a failed scan."""
         try:
